requirements-llm.txt
fastapi
uvicorn[standard]
chromadb>=0.4.16
ctransformers>=0.2.25,<0.3.0
deepmerge>=1.1.0,<2.0.0
InstructorEmbedding>=1.0.1,<2.0.0
openai
pydantic>=2.0.0